00_Docs/Prerequisites/The_AI_Engineer_Onramp_Cohort_2/Session_02_Back_End_Web_App_Development_&_Deployment_of_LLM_Applications/TreatOrHell/pyproject.toml
[project]
name = "treat-or-hell"
version = "0.1.0"
description = "A FastAPI app that lets you chat with St. Nicholas, an Angel, or a Devil using the OpenAI Chat Completions API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
]
